+++ src/app/appContext.js
@@ -4,11 +4,10 @@ import { createDivAndSetAttribute } from '../domMethods.js'
 
 export function appPushNewElements (attr) {
     if (isStr(attr), attr.search(/^(\.|\#)/) + 1, attr.slice(1).search(/^[\_|(a-zA-Z)]/) + 1) {
-
         var getElement = createDivAndSetAttribute(attr);
 
         {
-          getElement.innerHTML += dialogTemplate();
+          getElement.innerHTML = dialogTemplate.replace(/(^|\n)\s*/g,'');
           document.body.appendChild(getElement);
         }