+++ src/compatiblity.js
@@ -2,7 +2,7 @@ import { isExist } from './staticMethods.js'
 
 // validate style exist
 export const validateBrowserCompatiblityAnimationEvent = (el, eventObjectName) => {
-    for(var k in eventObjectName) {
+    for(const k in eventObjectName) {
         if(isExist(el.style[k])) {
             return eventObjectName[k]
         }